crawler-project/requirements.txt
cryptography
python-dateutil
python-dotenv
Scrapy